Recruitment-Automation-System/requirements.txt
streamlit>=1.28.1
openai>=1.3.9
python-dotenv>=1.0.0
pydantic>=2.5.0
PyPDF2>=3.0.1
python-docx>=0.8.11
spacy>=3.7.2
scikit-learn>=1.3.2
pandas>=2.1.3
numpy>=1.26.2
requests>=2.31.0
